refactor: address code assessment findings
Issue #2188 — Deep traversal: replace 4-level `_ctrl.state.total_linear_momentum.length()` in metrics_panel.py with new public `controller.total_momentum_magnitude()` method; replace 5-deep `Path(__file__).parent...parent` chain in env_utils.py with named `_REPO_ROOT_FALLBACK = Path(__file__).parents[4]` constant.

Issue #2187 — Hard-coded secret patterns: annotate `self.encryption_password = ""` in folder_packer_pro/app.py with `# noqa: S105` + comment clarifying it is a runtime UI field, not a hard-coded credential.

Issue #2186 — Duplicated responsibility: rename `calculate_fitting_pressure_drop(k_factor, density, velocity)` in fitting_loss_coefficients.py to `apply_k_factor` (which operates on a single K-factor scalar) to distinguish it clearly from the engine-level `calculate_fitting_pressure_drop(fittings, density, velocity, ...)` in engine/fittings.py that handles a list of PipeFitting objects. Backward-compatible alias retained.

Issue #2185 — Oversized module: extract `GrangerCausalityResult` and `TransferEntropyResult` data classes from cross_correlation.py (~1055 lines) into a new `causality_types.py` module; cross_correlation.py imports and re-exports them for backward compatibility. Remove the ARCHITECTURE_DEBT comment from core/__init__.py that was tracking this debt.
