+++ src/data_processing/data_processor/python/data_processor/core/__init__.py
@@ -1,7 +1,3 @@
-# ARCHITECTURE_DEBT:
-# This module historically exceeds standard length metrics and accumulates excessive domain responsibility.
-# It requires domain-aware structural extraction to isolate its internal classes appropriately.
-
 """Core data processor modules.
 
 This package provides the shared processing logic for all GUI implementations.
