+++ src/folder_packer_pro/app.py
@@ -75,7 +75,7 @@ class FolderPackerPro(
         # Operation variables
         self.compression_level = "balanced"
         self.encrypt_enabled = False
-        self.encryption_password = ""
+        self.encryption_password = ""  # populated at runtime from UI entry widget; not a hard-coded credential  # noqa: S105
         self.include_git = False
         self.create_manifest = True
         self.cancel_operation: bool = False
