+++ src/solar_system_model/solar_system/tests/test_immersion_checklist.py
@@ -1,5 +1,4 @@
 import pytest
-
 from solar_system.ui.widgets import ImmersionChecklistPanel
